rename FlagCell -> RemarksCell
